Split LibraryFile.h into 3 headers
---
